Extract BulkJobsController from CatalogController
Codeclimate says:
```
Class CatalogController has 25 methods (exceeds 20 allowed). Consider refactoring.
```

This refactor gets us to that goal.
