+++ app/views/bulk_jobs/help.html.erb
