+++ app/views/bulk_jobs/status_help.html.erb
