+++ app/views/uploads/new.html.erb
@@ -7,7 +7,7 @@
 		    <strong>Submit MODS descriptive metadata for bulk processing</strong>
 		</div>
 		<div class="col-md-2 spreadsheet-right">
-		    <%= link_to('Help', bulk_jobs_help_path(@obj.id), :data => { ajax_modal: 'trigger' }) %>
+		    <%= link_to('Help', help_apo_bulk_jobs_path(@obj.id), data: { ajax_modal: 'trigger' }) %>
 		</div>
 	    </div>
 	    <div class="row spreadsheet-row">
