+++ config/routes.rb
@@ -30,13 +30,6 @@ Argo::Application.routes.draw do
     end
   end
 
-  get 'catalog/:id/bulk_jobs_index',     to: 'catalog#bulk_jobs_index',   as: 'bulk_jobs_index'
-  get 'catalog/:id/bulk_status_help',    to: 'catalog#bulk_status_help',  as: 'bulk_status_help'
-  get 'catalog/:id/:time/bulk_jobs_xml', to: 'catalog#bulk_jobs_xml',     as: 'bulk_jobs_xml'
-  get 'catalog/:id/:time/bulk_jobs_log', to: 'catalog#bulk_jobs_log',     as: 'bulk_jobs_log'
-  get 'catalog/:id/bulk_jobs_help',      to: 'catalog#bulk_jobs_help',    as: 'bulk_jobs_help'
-  get 'catalog/:id/:time/bulk_jobs_csv', to: 'catalog#bulk_jobs_csv',     as: 'bulk_jobs_csv'
-  delete 'catalog/:id/bulk_jobs_delete', to: 'catalog#bulk_jobs_delete',  as: 'bulk_jobs_delete'
   match 'catalog',      via: [:get, :post], to: redirect { |params, req| req.fullpath.sub(%r{^/catalog}, '/view') }, as: 'search_catalog_redirect'
   match 'catalog/*all', via: [:get, :post], to: redirect { |params, req| req.fullpath.sub(%r{^/catalog}, '/view') }, as: 'catalog_redirect'
 
@@ -81,11 +74,23 @@ Argo::Application.routes.draw do
     end
   end
 
+  resources :apos, only: [] do
+    resources :bulk_jobs, only: :index do
+      collection do
+        get 'status_help'
+        get 'help'
+        get ':time/log', action: :show, as: 'show'
+        delete '', action: :destroy
+      end
+    end
+  end
+
   resources :items do
     resources 'uploads', only: [:new, :create]
 
     get :register, on: :collection
     resource :content_type, only: [:show, :update]
+
     member do
       get 'purl_preview'
       get 'discoverable'
